+++ src/front/src/index.css
@@ -18,3 +18,7 @@ input:-webkit-autofill:active {
   -webkit-transition: background-color 9999s ease-out;
   -webkit-box-shadow: 0 0 0px 1000px white inset !important;
 }
+
+.pointer {
+  cursor: pointer;
+}
